+++ src/components/common/Navigation/index.ts
@@ -1,0 +1,3 @@
+export { default as BottomNavigation } from './BottomNavigation';
+export { default as BottomNavigationBase } from './BottomNavigationBase';
+export { default as TopNavigation } from './TopNavigation';
